Feat/#48 update job upload file request

+++ src/main/groovy/io/saagie/plugin/dataops/models/Resources.groovy
@@ -2,8 +2,8 @@ package io.saagie.plugin.dataops.models
 
 class Resources implements IMapable {
     Float cpu
-    Float disk
-    Float memory
+    Integer disk
+    Integer memory
 
     Map toMap() {
         if (cpu && disk && memory) {
